fix(swaps): type RestoreTransaction.Vout as uint
A Bitcoin output index is a uint32. `long` accepts values no transaction can
have and does not match how the rest of the SDK carries a vout — NBitcoin's
OutPoint.N is uint, and RestClientTransport casts the indexer's vout to uint
before use. Any consumer of this field would have to narrow it back.

+++ NArk.Swaps/Boltz/Models/Restore/RestoreResponse.cs
@@ -71,7 +71,7 @@ public record RestoreTransaction
     /// Index of the lockup output in the transaction.
     /// </summary>
     [JsonPropertyName("vout")]
-    public required long Vout { get; init; }
+    public required uint Vout { get; init; }
 }
 
 /// <summary>
